Unify Project and Administration UI #1912
