+++ frontend/src/locale/en.json
@@ -63,7 +63,7 @@
     "fleets": "Fleets",
     "project": "Project",
     "project_other": "Projects",
-    "administration": "Administration",
+    "general": "General",
     "users": "Users",
     "user_settings": "User settings",
     "account": "User",
