+++ frontend/src/pages/Fleets/List/index.tsx
@@ -1,65 +1,124 @@
-import React, { useCallback, useMemo } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { Button, FormField, Header, Pagination, SelectCSD, Table, Toggle } from 'components';
-import { useProjectDropdown } from 'layouts/AppLayout/hooks';
 
-import { useCollection } from 'hooks';
-import { useGetFleetsQuery } from 'services/fleet';
+import { DEFAULT_TABLE_PAGE_SIZE } from 'consts';
+import { useBreadcrumbs, useCollection } from 'hooks';
+import { ROUTES } from 'routes';
+import { useLazyGetPoolsInstancesQuery } from 'services/pool';
 
 import { useColumnsDefinitions, useEmptyMessages, useFilters } from './hooks';
 
-import { TFleetInstance } from './types';
-
 import styles from './styles.module.scss';
 
-export const FleetsList: React.FC = () => {
+export const FleetList: React.FC = () => {
     const { t } = useTranslation();
-    const { selectedProject } = useProjectDropdown();
+    const [data, setData] = useState<IInstanceListItem[]>([]);
+    const [pagesCount, setPagesCount] = useState<number>(1);
+    const [disabledNext, setDisabledNext] = useState(false);
 
-    const { data, isLoading } = useGetFleetsQuery({ projectName: selectedProject ?? '' });
+    useBreadcrumbs([
+        {
+            text: t('navigation.fleets'),
+            href: ROUTES.FLEETS.LIST,
+        },
+    ]);
+
+    const {
+        onlyActive,
+        setOnlyActive,
+        isDisabledClearFilter,
+        clearFilters,
+        projectOptions,
+        selectedProject,
+        setSelectedProject,
+    } = useFilters();
+
+    const [getPools, { isLoading, isFetching }] = useLazyGetPoolsInstancesQuery();
+    const isDisabledPagination = isLoading || isFetching || data.length === 0;
+
+    const getPoolsRequest = (params?: Partial<TPoolInstancesRequestParams>) => {
+        return getPools({
+            only_active: onlyActive,
+            project_name: selectedProject?.value,
+            limit: DEFAULT_TABLE_PAGE_SIZE,
+            ...params,
+        }).unwrap();
+    };
+
+    useEffect(() => {
+        getPoolsRequest().then((result) => {
+            setPagesCount(1);
+            setDisabledNext(false);
+            setData(result);
+        });
+    }, [onlyActive, selectedProject?.value]);
 
     const { columns } = useColumnsDefinitions();
     const { renderEmptyMessage, renderNoMatchMessage } = useEmptyMessages();
 
-    const {
-        fleetOptions,
-        selectedFleet,
-        setSelectedFleet,
-        setOnlyActive,
-        onlyActive,
-        isDisabledClearFilter,
-        clearFilters,
-        filteringFunction,
-    } = useFilters({ fleets: data });
+    const nextPage = async () => {
+        if (data.length === 0 || disabledNext) {
+            return;
+        }
 
-    const fleetInstances = useMemo<TFleetInstance[]>(() => {
-        if (!data) return [];
-
-        return data.reduce<TFleetInstance[]>((acc, item, index) => {
-            if (!item.instances.length) {
-                acc.push({ fleetName: item.name });
-
-                return acc;
-            }
-
-            item.instances.forEach((instance) => {
-                acc.push({ fleetName: item.name, ...instance });
+        try {
+            const result = await getPoolsRequest({
+                prev_created_at: data[data.length - 1].created,
+                prev_id: data[data.length - 1].id,
             });
 
-            return acc;
-        }, []);
-    }, [data]);
+            if (result.length > 0) {
+                setPagesCount((count) => count + 1);
+                setData(result);
+            } else {
+                setDisabledNext(true);
+            }
+        } catch (e) {
+            console.log(e);
+        }
+    };
 
-    const { items, collectionProps, paginationProps } = useCollection<TFleetInstance[]>(fleetInstances, {
+    const prevPage = async () => {
+        if (pagesCount === 1) {
+            return;
+        }
+
+        try {
+            const result = await getPoolsRequest({
+                prev_created_at: data[0].created,
+                prev_id: data[0].id,
+                ascending: true,
+            });
+
+            setDisabledNext(false);
+
+            if (result.length > 0) {
+                setPagesCount((count) => count - 1);
+                setData(result);
+            } else {
+                setPagesCount(1);
+            }
+        } catch (e) {
+            console.log(e);
+        }
+    };
+
+    const { items, collectionProps } = useCollection<IInstanceListItem>(data, {
         filtering: {
             empty: renderEmptyMessage(),
             noMatch: renderNoMatchMessage(),
-            filteringFunction,
         },
         pagination: { pageSize: 20 },
         selection: {},
     });
+
+    const renderCounter = () => {
+        if (!data?.length) return '';
+
+        return `(${data.length})`;
+    };
 
     return (
         <Table
@@ -67,22 +126,26 @@ export const FleetsList: React.FC = () => {
             variant="full-page"
             columnDefinitions={columns}
             items={items}
-            loading={isLoading}
+            loading={isLoading || isFetching}
             loadingText={t('common.loading')}
             stickyHeader={true}
-            header={<Header variant="awsui-h1-sticky">{t('navigation.fleets')}</Header>}
+            header={
+                <Header variant="awsui-h1-sticky" counter={renderCounter()}>
+                    {t('navigation.fleets')}
+                </Header>
+            }
             filter={
                 <div className={styles.filters}>
                     <div className={styles.select}>
-                        <FormField label={t('fleets.fleet')}>
+                        <FormField label={t('projects.run.project')}>
                             <SelectCSD
-                                disabled={!fleetOptions?.length}
-                                options={fleetOptions}
-                                selectedOption={selectedFleet}
+                                disabled={!projectOptions?.length}
+                                options={projectOptions}
+                                selectedOption={selectedProject}
                                 onChange={(event) => {
-                                    setSelectedFleet(event.detail.selectedOption);
+                                    setSelectedProject(event.detail.selectedOption);
                                 }}
-                                placeholder={t('fleets.fleet_placeholder')}
+                                placeholder={t('projects.run.project_placeholder')}
                                 expandToViewport={true}
                                 filteringType="auto"
                             />
@@ -90,11 +153,7 @@ export const FleetsList: React.FC = () => {
                     </div>
 
                     <div className={styles.activeOnly}>
-                        <Toggle
-                            disabled={!fleetInstances.length}
-                            onChange={({ detail }) => setOnlyActive(detail.checked)}
-                            checked={onlyActive}
-                        >
+                        <Toggle onChange={({ detail }) => setOnlyActive(detail.checked)} checked={onlyActive}>
                             {t('fleets.active_only')}
                         </Toggle>
                     </div>
@@ -106,7 +165,16 @@ export const FleetsList: React.FC = () => {
                     </div>
                 </div>
             }
-            pagination={<Pagination {...paginationProps} disabled={isLoading} />}
+            pagination={
+                <Pagination
+                    currentPageIndex={pagesCount}
+                    pagesCount={pagesCount}
+                    openEnd={!disabledNext}
+                    disabled={isDisabledPagination}
+                    onPreviousPageClick={prevPage}
+                    onNextPageClick={nextPage}
+                />
+            }
         />
     );
 };
