--- frontend/src/pages/Fleets/List/types.ts
@@ -1,3 +1,0 @@
-export type TPoolWithProject = IPoolListItem & { project_name: string };
-
-export type TFleetInstance = Partial<IInstance> & { fleetName: IFleet['name'] };
