+++ frontend/src/pages/Models/List/index.tsx
@@ -2,30 +2,43 @@ import React, { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { sortBy as _sortBy } from 'lodash';
 
-import { Header, Pagination, Table } from 'components';
-import { useProjectDropdown } from 'layouts/AppLayout/hooks';
+import { Button, FormField, Header, Pagination, SelectCSD, Table } from 'components';
 
-import { useCollection } from 'hooks';
+import { useBreadcrumbs, useCollection } from 'hooks';
 import { getExtendedModelFromRun } from 'libs/run';
+import { ROUTES } from 'routes';
 import { useGetRunsQuery } from 'services/run';
 
+import { unfinishedRuns } from 'pages/Runs/constants';
+
 import { useModelListPreferences } from './Preferences/useModelListPreferences';
-import { unfinishedRuns } from '../../Runs/constants';
-import { useColumnsDefinitions, useEmptyMessages } from './hooks';
+import { useColumnsDefinitions, useEmptyMessages, useFilters } from './hooks';
 import { Preferences } from './Preferences';
 
 import { IModelExtended } from './types';
 
+import styles from './styles.module.scss';
+
 export const List: React.FC = () => {
     const { t } = useTranslation();
-    const { selectedProject } = useProjectDropdown();
 
-    const { data, isLoading } = useGetRunsQuery();
+    const { projectOptions, selectedProject, setSelectedProject, clearSelected } = useFilters({
+        projectSearchKey: 'project',
+    });
+
+    const { data, isLoading } = useGetRunsQuery({
+        project_name: selectedProject?.value,
+    });
+
+    useBreadcrumbs([
+        {
+            text: t('navigation.models'),
+            href: ROUTES.PROJECT.LIST,
+        },
+    ]);
 
     const { columns } = useColumnsDefinitions();
     const [preferences] = useModelListPreferences();
-
-    const { renderEmptyMessage, renderNoMatchMessage } = useEmptyMessages();
 
     const sortedData = useMemo<IModelExtended[]>(() => {
         if (!data) return [];
@@ -44,14 +57,21 @@ export const List: React.FC = () => {
         );
     }, [data]);
 
+    const clearFilter = () => {
+        clearSelected();
+    };
+
+    const isDisabledClearFilter = !selectedProject;
+
+    const { renderEmptyMessage, renderNoMatchMessage } = useEmptyMessages({
+        clearFilter,
+        isDisabledClearFilter,
+    });
+
     const { items, collectionProps, paginationProps } = useCollection<IModelExtended>(sortedData ?? [], {
         filtering: {
             empty: renderEmptyMessage(),
             noMatch: renderNoMatchMessage(),
-
-            filteringFunction: (modelItem) => {
-                return !(selectedProject && modelItem.project_name !== selectedProject);
-            },
         },
         pagination: { pageSize: 20 },
         selection: {},
@@ -68,6 +88,31 @@ export const List: React.FC = () => {
             stickyHeader={true}
             columnDisplay={preferences.contentDisplay}
             header={<Header variant="awsui-h1-sticky">{t('navigation.models')}</Header>}
+            filter={
+                <div className={styles.selectFilters}>
+                    <div className={styles.select}>
+                        <FormField label={t('projects.run.project')}>
+                            <SelectCSD
+                                disabled={!projectOptions?.length}
+                                options={projectOptions}
+                                selectedOption={selectedProject}
+                                onChange={(event) => {
+                                    setSelectedProject(event.detail.selectedOption);
+                                }}
+                                placeholder={t('projects.run.project_placeholder')}
+                                expandToViewport={true}
+                                filteringType="auto"
+                            />
+                        </FormField>
+                    </div>
+
+                    <div className={styles.clear}>
+                        <Button disabled={isDisabledClearFilter} formAction="none" onClick={clearSelected}>
+                            {t('common.clearFilter')}
+                        </Button>
+                    </div>
+                </div>
+            }
             pagination={<Pagination {...paginationProps} disabled={isLoading} />}
             preferences={<Preferences />}
         />
