+++ frontend/src/pages/Runs/index.ts
@@ -1,5 +1,3 @@
-export { List } from './MainList';
-export { AdministrationList } from './AdministrationList';
 export { RunList } from './List';
 export { RunDetails } from './Details';
 export { Logs } from './Details/Logs';
