--- frontend/src/pages/Volumes/AdministrationList/index.tsx
@@ -1,96 +1,0 @@
-import React from 'react';
-import { useTranslation } from 'react-i18next';
-
-import { Button, FormField, Header, Pagination, SelectCSD, Table, Toggle } from 'components';
-
-import { useCollection } from 'hooks';
-
-import { useFilters, useVolumesData, useVolumesTableEmptyMessages } from '../List/hooks';
-import { useColumnsDefinitions } from './hooks';
-
-import styles from '../List/styles.module.scss';
-
-export const AdministrationVolumeList: React.FC = () => {
-    const { t } = useTranslation();
-    const { renderEmptyMessage, renderNoMatchMessage } = useVolumesTableEmptyMessages();
-
-    const {
-        onlyActive,
-        setOnlyActive,
-        isDisabledClearFilter,
-        clearFilters,
-        projectOptions,
-        selectedProject,
-        setSelectedProject,
-    } = useFilters();
-
-    const { data, isLoading, pagesCount, disabledNext, prevPage, nextPage } = useVolumesData({
-        project_name: selectedProject?.value ?? undefined,
-        only_active: onlyActive,
-    });
-
-    const { columns } = useColumnsDefinitions();
-
-    const { items, actions, collectionProps } = useCollection(data, {
-        filtering: {
-            empty: renderEmptyMessage(),
-            noMatch: renderNoMatchMessage(() => actions.setFiltering('')),
-        },
-        pagination: { pageSize: 20 },
-        selection: {},
-    });
-
-    return (
-        <Table
-            {...collectionProps}
-            variant="full-page"
-            columnDefinitions={columns}
-            items={items}
-            loading={isLoading}
-            loadingText={t('common.loading')}
-            stickyHeader={true}
-            header={<Header>{t('volume.volumes')}</Header>}
-            filter={
-                <div className={styles.filters}>
-                    <div className={styles.select}>
-                        <FormField label={t('projects.run.project')}>
-                            <SelectCSD
-                                disabled={!projectOptions?.length}
-                                options={projectOptions}
-                                selectedOption={selectedProject}
-                                onChange={(event) => {
-                                    setSelectedProject(event.detail.selectedOption);
-                                }}
-                                placeholder={t('projects.run.project_placeholder')}
-                                expandToViewport={true}
-                                filteringType="auto"
-                            />
-                        </FormField>
-                    </div>
-
-                    <div className={styles.activeOnly}>
-                        <Toggle onChange={({ detail }) => setOnlyActive(detail.checked)} checked={onlyActive}>
-                            {t('volume.active_only')}
-                        </Toggle>
-                    </div>
-
-                    <div className={styles.clear}>
-                        <Button formAction="none" onClick={clearFilters} disabled={isDisabledClearFilter}>
-                            {t('common.clearFilter')}
-                        </Button>
-                    </div>
-                </div>
-            }
-            pagination={
-                <Pagination
-                    currentPageIndex={pagesCount}
-                    pagesCount={pagesCount}
-                    openEnd={!disabledNext}
-                    disabled={isLoading || data.length === 0}
-                    onPreviousPageClick={prevPage}
-                    onNextPageClick={nextPage}
-                />
-            }
-        />
-    );
-};
