+++ frontend/src/pages/Volumes/List/index.tsx
@@ -1,28 +1,42 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { Button, Header, Pagination, Table, Toggle } from 'components';
-import { useProjectDropdown } from 'layouts/AppLayout/hooks';
+import { Button, FormField, Header, Pagination, SelectCSD, Table, Toggle } from 'components';
 
-import { useCollection } from 'hooks';
+import { useBreadcrumbs, useCollection } from 'hooks';
+import { ROUTES } from 'routes';
 
 import { useColumnsDefinitions, useFilters, useVolumesData, useVolumesTableEmptyMessages } from './hooks';
 
-import styles from '../../Fleets/List/styles.module.scss';
+import styles from './styles.module.scss';
 
 export const VolumeList: React.FC = () => {
     const { t } = useTranslation();
     const { renderEmptyMessage, renderNoMatchMessage } = useVolumesTableEmptyMessages();
-    const { selectedProject } = useProjectDropdown();
 
-    const { onlyActive, setOnlyActive, isDisabledClearFilter, clearFilters } = useFilters();
+    const {
+        onlyActive,
+        setOnlyActive,
+        isDisabledClearFilter,
+        clearFilters,
+        projectOptions,
+        selectedProject,
+        setSelectedProject,
+    } = useFilters();
 
     const { data, isLoading, pagesCount, disabledNext, prevPage, nextPage } = useVolumesData({
-        project_name: selectedProject ?? undefined,
+        project_name: selectedProject?.value ?? undefined,
         only_active: onlyActive,
     });
 
     const { columns } = useColumnsDefinitions();
+
+    useBreadcrumbs([
+        {
+            text: t('volume.volumes'),
+            href: ROUTES.VOLUMES.LIST,
+        },
+    ]);
 
     const { items, actions, collectionProps } = useCollection(data, {
         filtering: {
@@ -45,6 +59,22 @@ export const VolumeList: React.FC = () => {
             header={<Header>{t('volume.volumes')}</Header>}
             filter={
                 <div className={styles.filters}>
+                    <div className={styles.select}>
+                        <FormField label={t('projects.run.project')}>
+                            <SelectCSD
+                                disabled={!projectOptions?.length}
+                                options={projectOptions}
+                                selectedOption={selectedProject}
+                                onChange={(event) => {
+                                    setSelectedProject(event.detail.selectedOption);
+                                }}
+                                placeholder={t('projects.run.project_placeholder')}
+                                expandToViewport={true}
+                                filteringType="auto"
+                            />
+                        </FormField>
+                    </div>
+
                     <div className={styles.activeOnly}>
                         <Toggle onChange={({ detail }) => setOnlyActive(detail.checked)} checked={onlyActive}>
                             {t('volume.active_only')}
