+++ frontend/src/pages/Volumes/index.ts
@@ -1,2 +1,1 @@
 export { VolumeList } from './List';
-export { AdministrationVolumeList } from './AdministrationList';
